debhelper: fix conffiles corner case
Debian records the files installed in /etc in a metadata file called
conffiles. It is generated on every build and it does not prune
duplicates.
The debhelper copies the one from the unsigned package to the signed
package, causing a duplicate to appear and a Lintian error to be
raised. Clean them up to fix it.