debhelper: restrict wildcard package unpacking
Generated debug symbols package append -dbg or -dbgsym to a package name
so a wildcard of foo*.deb will match both foo_3.deb and foo-dbg_3.deb.
Extend the wildcard to foo_*.deb so that it matches until the version
separator and avoids accidentally including other packages that include
the same name prefix, like the aforementioned -dbg or -dbgsym.