Compress kernel modules in batch and in parallel
currently we're sequentially compressing the modules using
xz -9. In testing against the kernel 5.8 in Tumbleweed, there
was very little compression size difference between xz -9 and xz,
however almost a 40% runtime performance improvement.

Also running in 4 jobs in parallel (instead of using xz -threads)
improves speed quite a bit. Last but not least wrapping
the invocations via "xargs -t" ensures that there is a constant
stream of logging output that prevents the build service
from killing the job as it is considered stalled.

Signed-off-by: Dirk Mueller <dirk@dmllr.de>